recipeの投稿、更新、削除にadmin権限付与

+++ api/src/recipes/recipes.service.ts
@@ -1,10 +1,15 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { RecipeRepository } from './recipe.repository';
 import { CreateRecipeDto } from './dto/create-recipe.dto';
 import { Recipe } from './recipe.entity';
 import { GetRecipesFilterDto } from './dto/get-recipes.dto';
 import { MyKnownMessage } from '../message.interface';
+import { User } from '../users/user.entity';
 
 @Injectable()
 export class RecipesService {
@@ -28,14 +33,21 @@ export class RecipesService {
     return found;
   }
 
-  async createRecipe(createRecipeDto: CreateRecipeDto): Promise<Recipe> {
-    return this.recipeRepository.createRecipe(createRecipeDto);
+  async createRecipe(
+    createRecipeDto: CreateRecipeDto,
+    user: User,
+  ): Promise<Recipe> {
+    return this.recipeRepository.createRecipe(createRecipeDto, user);
   }
 
   async updateRecipe(
     id: number,
     createRecipeDto: CreateRecipeDto,
+    user: User,
   ): Promise<Recipe> {
+    if (user.role !== 'admin') {
+      throw new UnauthorizedException('権限がありません');
+    }
     const found = await this.getRecipeById(id);
     const { name, time, remarks, image } = createRecipeDto;
 
@@ -48,7 +60,10 @@ export class RecipesService {
     return found;
   }
 
-  async deleteRecipe(id: number): Promise<MyKnownMessage> {
+  async deleteRecipe(id: number, user: User): Promise<MyKnownMessage> {
+    if (user.role !== 'admin') {
+      throw new UnauthorizedException('権限がありません');
+    }
     const result = await this.recipeRepository.delete({ id });
 
     if (result.affected === 0) {
